refactor: consolidate the cipher-inversion engine and move partial-cipher to editor + doctests
